feat: open the focused active session conversation directly
Clicking an already-focused Active Session card now opens the AI
Conversation viewer at its latest input instead of expanding an inline
outline rail, whose prompt outline duplicated the viewer's richer
searchable outline. The viewer's Outline and Comments header buttons
merge into one Sidebar toggle and the side panel starts closed.
