+++ docs/testing/main-capability-coverage.json
@@ -1037,17 +1037,14 @@
             "behaviors": [
                 "SESSION-AI-SESSION-CONVERSATION-ADAPTER-001",
                 "SESSION-CLAUDE-SUBAGENT-LIFECYCLE-001",
-                "WEBVIEW-AI-SESSION-CONVERSATION-OUTLINE-001",
+                "ACTIVE-SESSION-CONVERSATION-OPEN-001",
                 "WEBVIEW-AI-SESSION-CONVERSATION-VIEWER-001",
                 "SECURITY-AI-SESSION-CONVERSATION-SOURCE-001",
-                "ACTIVE-SESSION-CONVERSATION-RESTORE-001",
-                "SESSION-CONVERSATION-LOADING-001",
                 "CONVERSATION-VIEWER-LOADING-001",
                 "CONVERSATION-PROTOCOL-VALIDATOR-001",
                 "CONVERSATION-REFRESH-PERFORMANCE-001",
                 "CONVERSATION-LARGE-SESSION-PERFORMANCE-001",
                 "CONVERSATION-COMMENTS-DOM-STABILITY-001",
-                "CONVERSATION-OUTLINE-DELIVERY-001",
                 "CONVERSATION-VIEWER-USER-EMPHASIS-001",
                 "CONVERSATION-COMMENTS-001",
                 "CONVERSATION-COMMENTS-SUBMIT-001",
