+++ media/styles.scss
@@ -3288,10 +3288,7 @@ body.steward-sidebar {
             gap: 2px;
             margin-top: 5px;
             margin-left: 0;
-            height: calc(
-                var(--steward-ai-session-list-max-height, calc(3 * 42px + 2 * 2px))
-                + var(--steward-ai-session-expanded-extra-height, 0px)
-            );
+            height: var(--steward-ai-session-list-max-height, calc(3 * 42px + 2 * 2px));
             overflow-y: auto;
             overflow-x: hidden;
             padding-right: 2px;
@@ -3389,11 +3386,11 @@ body.steward-sidebar {
             }
         }
 
-        .active-ai-session-row[data-session-focused] .ai-session-primary-action[aria-controls] {
+        .active-ai-session-row[data-session-focused] .ai-session-primary-action {
             grid-template-columns: 28px minmax(0, 1fr) 14px;
         }
 
-        .ai-session-conversation-chevron {
+        .ai-session-open-conversation-hint {
             display: inline-flex;
             align-items: center;
             justify-content: center;
@@ -3401,186 +3398,6 @@ body.steward-sidebar {
             color: var(--vscode-descriptionForeground);
             font-size: 18px;
             line-height: 1;
-            transform: rotate(0);
-            transform-origin: center;
-            transition: transform $defaultTransition;
-        }
-
-        .ai-session-conversation-chevron[data-expanded] {
-            transform: rotate(90deg);
-        }
-
-        .active-ai-session-row[data-conversation-expanded] {
-            display: grid;
-            grid-template-rows: auto minmax(0, auto);
-
-            .codex-session-actions {
-                top: 20px;
-            }
-        }
-
-        .ai-session-conversation-panel {
-            display: grid;
-            grid-template-rows: auto auto minmax(0, auto);
-            min-height: 0;
-            margin: 0 7px 7px;
-            padding: 6px 7px 7px;
-            border-top: 1px solid var(--vscode-panel-border);
-            overflow: hidden;
-
-            &[hidden] {
-                display: none;
-            }
-
-            > header {
-                display: flex;
-                align-items: center;
-                justify-content: space-between;
-                min-height: 22px;
-                color: var(--vscode-descriptionForeground);
-                font-size: 10px;
-                font-weight: 650;
-                letter-spacing: .04em;
-            }
-        }
-
-        .ai-session-conversation-loading {
-            display: grid;
-            gap: 5px;
-            align-content: start;
-            min-height: 24px;
-            padding: 8px 0;
-            box-sizing: border-box;
-            color: var(--vscode-descriptionForeground);
-            font-size: 10px;
-
-            &[data-state="loading"] {
-                min-height: 72px;
-            }
-
-            &[data-state="partial"] {
-                padding: 2px 0 5px;
-                color: var(--vscode-editorWarning-foreground, var(--vscode-descriptionForeground));
-            }
-        }
-
-        .ai-session-conversation-loading[hidden],
-        .ai-session-conversation-rail[hidden] {
-            display: none;
-        }
-
-        .ai-session-conversation-rail {
-            display: grid;
-            align-content: start;
-            gap: 0;
-            min-height: 0;
-            height: var(
-                --steward-ai-session-conversation-rail-height,
-                auto
-            );
-            max-height: var(
-                --steward-ai-session-conversation-rail-height,
-                168px
-            );
-            overflow-x: hidden;
-            overflow-y: auto;
-            overscroll-behavior: contain;
-        }
-
-        .ai-session-conversation-marker {
-            position: relative;
-            display: grid;
-            grid-template-columns: 14px minmax(0, 1fr);
-            align-items: center;
-            gap: 8px;
-            width: 100%;
-            height: 28px;
-            min-height: 28px;
-            padding: 0 6px 0 0;
-            border: 0;
-            outline: none;
-            overflow: hidden;
-            color: var(--vscode-descriptionForeground);
-            background: transparent;
-            font: inherit;
-            font-size: 11px;
-            line-height: 1.3;
-            text-align: left;
-            cursor: pointer;
-            box-sizing: border-box;
-
-            &:hover .ai-session-conversation-marker-stroke,
-            &:focus-visible .ai-session-conversation-marker-stroke,
-            &[aria-selected="true"] .ai-session-conversation-marker-stroke {
-                opacity: .78;
-            }
-
-            &:focus-visible {
-                outline: 1px solid var(--vscode-focusBorder);
-                outline-offset: -2px;
-            }
-
-            &[data-latest] {
-                color: var(--vscode-foreground);
-
-                .ai-session-conversation-marker-stroke {
-                    opacity: .9;
-                }
-            }
-
-            &[data-current]::after {
-                content: "";
-                position: absolute;
-                top: 50%;
-                right: 1px;
-                width: 4px;
-                height: 4px;
-                border-radius: 50%;
-                background: var(--steward-ai-accent, var(--vscode-focusBorder));
-                opacity: .65;
-                transform: translateY(-50%);
-            }
-        }
-
-        .ai-session-conversation-marker-stroke {
-            width: 14px;
-            height: 2px;
-            border-radius: 999px;
-            background: var(--steward-ai-accent, var(--vscode-focusBorder));
-            opacity: .48;
-        }
-
-        .ai-session-conversation-marker-preview {
-            min-width: 0;
-            overflow: hidden;
-            text-overflow: ellipsis;
-            white-space: nowrap;
-        }
-
-        .ai-session-conversation-state-hint {
-            color: var(--vscode-descriptionForeground);
-            line-height: 1.35;
-        }
-
-        .ai-session-conversation-retry {
-            justify-self: start;
-            min-height: 24px;
-            padding: 2px 8px;
-            border: 1px solid var(--vscode-button-border, transparent);
-            border-radius: 4px;
-            color: var(--vscode-button-foreground);
-            background: var(--vscode-button-background);
-            font: inherit;
-            cursor: pointer;
-
-            &:hover:not(:disabled) {
-                background: var(--vscode-button-hoverBackground);
-            }
-
-            &:disabled {
-                cursor: default;
-                opacity: .55;
-            }
         }
 
         .codex-session-row[data-session-provider="kimi"] {
