+++ src/webview/webviewProjectScripts.js
@@ -46,8 +46,13 @@ function getAiSessionCardActivation(target, projectId) {
             return {
                 handled: true,
                 sessionRow: activationSessionRow,
-                message: null,
-                toggleConversation: true,
+                message: {
+                    type: 'open-active-ai-session-conversation',
+                    version: 1,
+                    projectId: projectId,
+                    provider: provider,
+                    sessionId: sessionId,
+                },
             };
         }
         return {
@@ -74,371 +79,6 @@ function getAiSessionCardActivation(target, projectId) {
             sessionId: sessionId,
         },
     };
-}
-
-var expandedActiveAiSessionConversationKey = null;
-var activeAiSessionConversationSubscription = null;
-var activeAiSessionConversationRequestId = 0;
-var activeAiSessionConversationGeneration = 0;
-var activeAiSessionConversationResizeObserver = null;
-var activeAiSessionConversationMutationObserver = null;
-var activeAiSessionConversationResizeFallbackInstalled = false;
-var activeAiSessionConversationRetryTimer = null;
-var activeAiSessionConversationRetryDeadline = 0;
-
-var ACTIVE_AI_SESSION_CONVERSATION_PREVIEW_GRAPHEMES = 160;
-var ACTIVE_AI_SESSION_CONVERSATION_MAX_INTERACTIONS = 2000;
-var ACTIVE_AI_SESSION_CONVERSATION_MAX_MESSAGE_GRAPHEMES = 64 * 1000;
-var ACTIVE_AI_SESSION_CONVERSATION_MAX_PREVIEW_CODE_UNITS = 4096;
-var ACTIVE_AI_SESSION_CONVERSATION_MAX_RETRY_AFTER_MS = 60 * 1000;
-var ACTIVE_AI_SESSION_CONVERSATION_RESPONSE_STATES = [
-    'complete', 'inProgress', 'interrupted', 'unknown',
-];
-var ACTIVE_AI_SESSION_CONVERSATION_ERROR_CODES = [
-    'unavailable', 'staleRevision', 'unsupportedVersion', 'tooLarge', 'timeout',
-];
-var ACTIVE_AI_SESSION_CONVERSATION_ERROR_REASONS = [
-    'missingSource', 'updateCodex', 'unsupportedCodexProtocol',
-    'reconnectingCodex', 'codexRetryExhausted',
-];
-var activeAiSessionConversationSegmenter = null;
-
-function isPlainConversationObject(value) {
-    return !!value && typeof value === 'object' && !Array.isArray(value);
-}
-
-function hasExactConversationKeys(value, required, optional) {
-    if (!isPlainConversationObject(value)) return false;
-    var requiredKeys = required || [];
-    var allowed = new Set(requiredKeys.concat(optional || []));
-    var keys = Object.keys(value);
-    return requiredKeys.every(key => Object.prototype.hasOwnProperty.call(value, key))
-        && keys.every(key => allowed.has(key));
-}
-
-function isBoundedConversationIdentity(value) {
-    return typeof value === 'string'
-        && value.trim().length > 0
-        && value.length <= 1024;
-}
-
-function conversationGraphemes(value) {
-    value = typeof value === 'string' ? value : '';
-    if (typeof Intl !== 'undefined' && typeof Intl.Segmenter === 'function') {
-        activeAiSessionConversationSegmenter =
-            activeAiSessionConversationSegmenter
-            || new Intl.Segmenter(undefined, { granularity: 'grapheme' });
-        return Array.from(
-            activeAiSessionConversationSegmenter.segment(value),
-            segment => segment.segment
-        );
-    }
-    return Array.from(value);
-}
-
-function truncateConversationPreview(value) {
-    if (value.length <= ACTIVE_AI_SESSION_CONVERSATION_PREVIEW_GRAPHEMES) {
-        return value;
-    }
-    return conversationGraphemes(value)
-        .slice(0, ACTIVE_AI_SESSION_CONVERSATION_PREVIEW_GRAPHEMES)
-        .join('');
-}
-
-function isValidConversationSummary(summary) {
-    if (!hasExactConversationKeys(
-        summary,
-        ['id', 'userPreview', 'userGraphemeCount', 'responseState'],
-        ['providerTurnId', 'timestamp']
-    )
-        || !isBoundedConversationIdentity(summary.id)
-        || typeof summary.userPreview !== 'string'
-        || summary.userPreview.length
-            > ACTIVE_AI_SESSION_CONVERSATION_MAX_PREVIEW_CODE_UNITS
-        || conversationGraphemes(summary.userPreview).length
-            > ACTIVE_AI_SESSION_CONVERSATION_PREVIEW_GRAPHEMES
-        || !Number.isSafeInteger(summary.userGraphemeCount)
-        || summary.userGraphemeCount < 0
-        || summary.userGraphemeCount
-            > ACTIVE_AI_SESSION_CONVERSATION_MAX_MESSAGE_GRAPHEMES
-        || !ACTIVE_AI_SESSION_CONVERSATION_RESPONSE_STATES.includes(
-            summary.responseState
-        )
-        || (summary.providerTurnId !== undefined
-            && !isBoundedConversationIdentity(summary.providerTurnId))
-        || (summary.timestamp !== undefined
-            && (!Number.isFinite(summary.timestamp) || summary.timestamp < 0))) {
-        return false;
-    }
-    return true;
-}
-
-function isValidConversationOutline(outline, state) {
-    if (!hasExactConversationKeys(
-        outline,
-        [
-            'provider', 'sessionId', 'sourceRevision',
-            'totalInteractions', 'partial', 'interactions',
-        ],
-        []
-    )
-        || outline.provider !== state.provider
-        || outline.sessionId !== state.sessionId
-        || !isBoundedConversationIdentity(outline.sourceRevision)
-        || !Number.isSafeInteger(outline.totalInteractions)
-        || outline.totalInteractions < 0
-        || typeof outline.partial !== 'boolean'
-        || !Array.isArray(outline.interactions)
-        || outline.interactions.length > ACTIVE_AI_SESSION_CONVERSATION_MAX_INTERACTIONS
-        || outline.totalInteractions < outline.interactions.length
-        || !outline.interactions.every(isValidConversationSummary)) {
-        return false;
-    }
-    var ids = new Set(outline.interactions.map(summary => summary.id));
-    return ids.size === outline.interactions.length;
-}
-
-function isValidConversationError(error, state) {
-    if (!hasExactConversationKeys(error, ['code'], ['reason', 'retryAfterMs'])
-        || !ACTIVE_AI_SESSION_CONVERSATION_ERROR_CODES.includes(error.code)
-        || (error.reason !== undefined
-            && !ACTIVE_AI_SESSION_CONVERSATION_ERROR_REASONS.includes(error.reason))) {
-        return false;
-    }
-    if (error.reason === undefined) {
-        return error.retryAfterMs === undefined;
-    }
-    if ([
-        'updateCodex', 'unsupportedCodexProtocol',
-        'reconnectingCodex', 'codexRetryExhausted',
-    ].includes(error.reason)
-        && state.provider !== 'codex') {
-        return false;
-    }
-    if (error.reason === 'codexRetryExhausted') {
-        return error.code === 'unavailable'
-            && Number.isSafeInteger(error.retryAfterMs)
-            && error.retryAfterMs > 0
-            && error.retryAfterMs
-                <= ACTIVE_AI_SESSION_CONVERSATION_MAX_RETRY_AFTER_MS;
-    }
-    if (error.retryAfterMs !== undefined) return false;
-    if (error.reason === 'unsupportedCodexProtocol') {
-        return error.code === 'unsupportedVersion';
-    }
-    return error.code === 'unavailable';
-}
-
-function clearActiveAiSessionConversationRetryTimer() {
-    if (activeAiSessionConversationRetryTimer !== null) {
-        window.clearTimeout(activeAiSessionConversationRetryTimer);
-        activeAiSessionConversationRetryTimer = null;
-    }
-    activeAiSessionConversationRetryDeadline = 0;
-}
-
-function clearConversationElement(element) {
-    if (!element) return;
-    while (element.firstChild) {
-        element.removeChild(element.firstChild);
-    }
-}
-
-function setActiveAiSessionConversationStatus(
-    row,
-    stateName,
-    text,
-    hint,
-    retryAfterMs,
-    retainedRetryDeadline
-) {
-    var panel = row && row.querySelector('[data-ai-session-conversation-panel]');
-    var status = panel && panel.querySelector('.ai-session-conversation-loading');
-    if (!status) return null;
-    clearConversationElement(status);
-    status.hidden = false;
-    status.setAttribute('data-ai-session-conversation-state', '');
-    status.setAttribute('data-state', stateName);
-    var textNode = document.createElement('span');
-    textNode.textContent = text;
-    status.appendChild(textNode);
-    if (hint) {
-        var hintNode = document.createElement('span');
-        hintNode.className = 'ai-session-conversation-state-hint';
-        hintNode.textContent = hint;
-        status.appendChild(hintNode);
-    }
-    if (stateName !== 'loading' && stateName !== 'empty'
-        && stateName !== 'partial') {
-        var retry = document.createElement('button');
-        retry.type = 'button';
-        retry.className = 'ai-session-conversation-retry';
-        retry.setAttribute('data-action', 'retry-ai-session-conversation');
-        retry.textContent = 'Retry';
-        var requestedDelay = Number.isSafeInteger(retryAfterMs) && retryAfterMs > 0
-            ? retryAfterMs
-            : 0;
-        var hasRetainedDeadline = Number.isFinite(retainedRetryDeadline);
-        var retryDeadline = hasRetainedDeadline
-            ? retainedRetryDeadline
-            : requestedDelay > 0
-                ? Date.now() + requestedDelay
-                : 0;
-        var delay = Math.max(0, retryDeadline - Date.now());
-        retry.disabled = delay > 0;
-        retry.addEventListener('click', event => {
-            event.preventDefault();
-            event.stopPropagation();
-            if (retry.disabled
-                || activeAiSessionConversationSubscription === null) return;
-            var currentRow = getCurrentActiveAiSessionConversationRow(
-                activeAiSessionConversationSubscription
-            );
-            if (currentRow === row) {
-                requestActiveAiSessionConversation(row);
-            }
-        });
-        status.appendChild(retry);
-        if (delay > 0) {
-            activeAiSessionConversationRetryDeadline = retryDeadline;
-            var retryState = activeAiSessionConversationSubscription;
-            var enableRetry = () => {
-                if (activeAiSessionConversationSubscription !== retryState
-                    || !retry.isConnected) {
-                    clearActiveAiSessionConversationRetryTimer();
-                    return;
-                }
-                var remaining = activeAiSessionConversationRetryDeadline - Date.now();
-                if (remaining > 0) {
-                    activeAiSessionConversationRetryTimer = window.setTimeout(
-                        enableRetry,
-                        remaining
-                    );
-                    return;
-                }
-                retry.disabled = false;
-                activeAiSessionConversationRetryTimer = null;
-                activeAiSessionConversationRetryDeadline = 0;
-                if (retryState.validatedRender?.kind === 'error') {
-                    retryState.validatedRender.retryDeadline = 0;
-                }
-            };
-            activeAiSessionConversationRetryTimer = window.setTimeout(
-                enableRetry,
-                delay
-            );
-        }
-    }
-    return status;
-}
-
-function prepareActiveAiSessionConversationLoading(row) {
-    clearActiveAiSessionConversationRetryTimer();
-    var panel = row && row.querySelector('[data-ai-session-conversation-panel]');
-    var rail = panel && panel.querySelector('[data-ai-session-conversation-rail]');
-    var count = panel && panel.querySelector('[data-ai-session-conversation-count]');
-    if (!panel || !rail || !count) return false;
-    clearConversationElement(rail);
-    rail.hidden = true;
-    count.textContent = '0';
-    setActiveAiSessionConversationStatus(
-        row,
-        'loading',
-        'Loading conversation…'
-    );
-    return true;
-}
-
-function getActiveAiSessionConversationTarget(row) {
-    if (!row || typeof row.closest !== 'function') return null;
-    var projectDiv = row.closest('.project[data-id]');
-    var projectId = projectDiv && projectDiv.getAttribute('data-id');
-    var provider = row.getAttribute('data-session-provider');
-    var sessionId = row.getAttribute('data-session-id');
-    if (!projectId || !sessionId
-        || (provider !== 'codex' && provider !== 'kimi' && provider !== 'claude')) {
-        return null;
-    }
-    return { projectId: projectId, provider: provider, sessionId: sessionId };
-}
-
-function getActiveAiSessionConversationKey(target) {
-    return target
-        ? JSON.stringify([target.projectId, target.provider, target.sessionId])
-        : null;
-}
-
-function nextActiveAiSessionConversationEnvelope(target) {
-    activeAiSessionConversationRequestId += 1;
-    activeAiSessionConversationGeneration += 1;
-    return {
-        version: 1,
-        requestId: activeAiSessionConversationRequestId,
-        subscriptionGeneration: activeAiSessionConversationGeneration,
-        projectId: target.projectId,
-        provider: target.provider,
-        sessionId: target.sessionId,
-    };
-}
-
-function requestActiveAiSessionConversation(row, restoreState) {
-    var target = getActiveAiSessionConversationTarget(row);
-    if (!target || !window.vscode || typeof window.vscode.postMessage !== 'function') {
-        return false;
-    }
-    prepareActiveAiSessionConversationLoading(row);
-    var envelope = nextActiveAiSessionConversationEnvelope(target);
-    activeAiSessionConversationSubscription = Object.assign({}, target, envelope, {
-        outlineRequestId: envelope.requestId,
-        requestId: activeAiSessionConversationRequestId,
-        expanded: true,
-        sourceRevision: null,
-        interactionIds: new Set(),
-        hasRenderedOutline: false,
-        validatedRender: null,
-        restoreState: restoreState || null,
-    });
-    window.vscode.postMessage(Object.assign({
-        type: 'request-ai-session-conversation-outline',
-    }, envelope));
-    return true;
-}
-
-function cancelActiveAiSessionConversation(target) {
-    target = target || activeAiSessionConversationSubscription;
-    clearActiveAiSessionConversationRetryTimer();
-    if (!target || !window.vscode || typeof window.vscode.postMessage !== 'function') {
-        activeAiSessionConversationSubscription = null;
-        expandedActiveAiSessionConversationKey = null;
-        disconnectActiveAiSessionConversationResizeObserver();
-        return false;
-    }
-    var envelope = nextActiveAiSessionConversationEnvelope(target);
-    window.vscode.postMessage(Object.assign({
-        type: 'cancel-ai-session-conversation',
-    }, envelope));
-    activeAiSessionConversationSubscription = null;
-    expandedActiveAiSessionConversationKey = null;
-    disconnectActiveAiSessionConversationResizeObserver();
-    return true;
-}
-
-function getCurrentActiveAiSessionConversationRow(state) {
-    if (!state
-        || typeof document === 'undefined'
-        || typeof document.querySelectorAll !== 'function') return null;
-    return Array.from(document.querySelectorAll(
-        '.workspace-card[data-current-workspace][data-id] '
-        + '.active-ai-session-row[data-conversation-expanded][data-session-focused]'
-    )).find(row => {
-        var target = getActiveAiSessionConversationTarget(row);
-        return target
-            && target.projectId === state.projectId
-            && target.provider === state.provider
-            && target.sessionId === state.sessionId
-            && getActiveAiSessionConversationKey(target)
-                === expandedActiveAiSessionConversationKey;
-    }) || null;
 }
 
 function parseAiSessionConversationFocusOrigin(message) {
@@ -483,26 +123,14 @@ function focusAiSessionConversationOrigin(message) {
     selectAiSessionTabDom(projectDiv, 'active');
     writeAiSessionTabState(window.vscode, origin.projectId, 'active');
     var row = Array.from(projectDiv.querySelectorAll(
-        '.active-ai-session-row[data-conversation-expanded]'
-        + '[data-session-focused][data-session-provider][data-session-id]'
+        '.active-ai-session-row[data-session-focused]'
+        + '[data-session-provider][data-session-id]'
     )).find(candidate =>
         candidate.getAttribute('data-session-provider') === origin.provider
         && candidate.getAttribute('data-session-id') === origin.sessionId
     );
     if (row) {
-        var marker = Array.from(row.querySelectorAll(
-            '[data-ai-session-conversation-marker][data-interaction-id]'
-        )).find(candidate =>
-            candidate.getAttribute('data-interaction-id')
-                === origin.interactionId
-        );
         var header = row.querySelector('.ai-session-primary-action');
-        if (marker && typeof marker.focus === 'function') {
-            marker.focus({ preventScroll: true });
-        }
-        if (marker && document.activeElement === marker) {
-            return true;
-        }
         if (header && typeof header.focus === 'function') {
             header.focus({ preventScroll: true });
         }
@@ -520,742 +148,6 @@ function focusAiSessionConversationOrigin(message) {
         return document.activeElement === activeTab;
     }
     return false;
-}
-
-function setConversationMarkerSelection(markers, index, shouldFocus) {
-    if (!markers.length) return;
-    var bounded = Math.max(0, Math.min(markers.length - 1, index));
-    markers.forEach((marker, markerIndex) => {
-        var selected = markerIndex === bounded;
-        marker.tabIndex = selected ? 0 : -1;
-        marker.setAttribute('aria-selected', selected ? 'true' : 'false');
-    });
-    if (shouldFocus) {
-        markers[bounded]?.focus({ preventScroll: true });
-    }
-}
-
-function focusConversationMarker(markers, index) {
-    setConversationMarkerSelection(markers, index, true);
-}
-
-function activateConversationMarker(marker, state, vscode) {
-    if (!marker || !state || !state.expanded || !state.sourceRevision
-        || state !== activeAiSessionConversationSubscription
-        || getCurrentActiveAiSessionConversationRow(state) === null
-        || !vscode || typeof vscode.postMessage !== 'function') return;
-    var interactionId = marker.getAttribute('data-interaction-id') || '';
-    if (!state.interactionIds.has(interactionId)) return;
-    activeAiSessionConversationRequestId += 1;
-    state.requestId = activeAiSessionConversationRequestId;
-    vscode.postMessage({
-        type: 'open-ai-session-conversation',
-        version: 1,
-        requestId: state.requestId,
-        subscriptionGeneration: state.subscriptionGeneration,
-        projectId: state.projectId,
-        provider: state.provider,
-        sessionId: state.sessionId,
-        interactionId: interactionId,
-        expectedRevision: state.sourceRevision,
-    });
-}
-
-function getConversationTimestampLabel(timestamp) {
-    if (timestamp === undefined) return 'Time unavailable';
-    try {
-        return new Date(timestamp).toISOString();
-    } catch (_error) {
-        return 'Time unavailable';
-    }
-}
-
-function getConversationAutoScrollThreshold(rail) {
-    var value = rail && rail.getAttribute('data-auto-scroll-threshold');
-    if (typeof value !== 'string'
-        || !/^(?:\d+|\d*\.\d+)$/.test(value.trim())) return null;
-    var threshold = Number(value);
-    return Number.isFinite(threshold) && threshold >= 0 ? threshold : null;
-}
-
-function getConversationMarkerKey(marker) {
-    return marker.getAttribute('data-interaction-id') || '';
-}
-
-function captureConversationRailState(rail) {
-    var endThreshold = getConversationAutoScrollThreshold(rail);
-    return window.__agentPivotScrollState.capture(rail, {
-        itemSelector:
-            '[data-ai-session-conversation-marker][data-interaction-id]',
-        getKey: getConversationMarkerKey,
-        endThreshold: endThreshold === null ? undefined : endThreshold,
-    });
-}
-
-function restoreConversationRailState(rail, anchor) {
-    return window.__agentPivotScrollState.restore(rail, anchor, {
-        itemSelector:
-            '[data-ai-session-conversation-marker][data-interaction-id]',
-        getKey: getConversationMarkerKey,
-        followEnd: true,
-    });
-}
-
-function revealConversationMarker(rail, marker) {
-    if (!rail || !marker || rail.scrollHeight <= rail.clientHeight) return;
-    var railRect = rail.getBoundingClientRect();
-    var markerRect = marker.getBoundingClientRect();
-    if (markerRect.bottom > railRect.bottom) {
-        rail.scrollTop += markerRect.bottom - railRect.bottom;
-    } else if (markerRect.top < railRect.top) {
-        rail.scrollTop -= railRect.top - markerRect.top;
-    }
-}
-
-function handleConversationMarkerKeydown(event, marker, state) {
-    var rail = marker && marker.closest('[data-ai-session-conversation-rail]');
-    var markers = rail ? Array.from(rail.querySelectorAll(
-        '[data-ai-session-conversation-marker]'
-    )) : [];
-    var index = markers.indexOf(marker);
-    if (index < 0) return;
-    var destination = null;
-    if (event.key === 'ArrowUp') destination = index - 1;
-    else if (event.key === 'ArrowDown') destination = index + 1;
-    else if (event.key === 'Home') destination = 0;
-    else if (event.key === 'End') destination = markers.length - 1;
-    else if (event.key === 'Enter') {
-        event.preventDefault();
-        event.stopPropagation();
-        activateConversationMarker(marker, state, window.vscode);
-        return;
-    } else {
-        return;
-    }
-    event.preventDefault();
-    event.stopPropagation();
-    focusConversationMarker(markers, destination);
-    revealConversationMarker(
-        rail,
-        markers[Math.max(0, Math.min(markers.length - 1, destination))]
-    );
-}
-
-function renderActiveAiSessionConversationOutline(row, state, outline) {
-    var panel = row && row.querySelector('[data-ai-session-conversation-panel]');
-    var rail = panel && panel.querySelector('[data-ai-session-conversation-rail]');
-    var count = panel && panel.querySelector('[data-ai-session-conversation-count]');
-    var status = panel && panel.querySelector('.ai-session-conversation-loading');
-    if (!panel || !rail || !count || !status) return false;
-
-    clearActiveAiSessionConversationRetryTimer();
-    var currentRailAnchor = captureConversationRailState(rail);
-    var focusedMarker = rail.querySelector(
-        '[data-ai-session-conversation-marker]:focus'
-    );
-    var focusedInteractionId = focusedMarker
-        ? focusedMarker.getAttribute('data-interaction-id') || ''
-        : '';
-    var restoreState = state.restoreState;
-    state.restoreState = null;
-
-    clearConversationElement(rail);
-    state.sourceRevision = outline.sourceRevision;
-    state.interactionIds = new Set(outline.interactions.map(summary => summary.id));
-    state.validatedRender = {
-        kind: 'outline',
-        outline: outline,
-    };
-    var retainedFocusedInteractionId = restoreState?.focusedInteractionId
-        || focusedInteractionId;
-    var selectedInteractionId = retainedFocusedInteractionId
-        || outline.interactions.at(-1)?.id
-        || '';
-
-    outline.interactions.forEach((summary, index) => {
-        var marker = document.createElement('button');
-        var preview = truncateConversationPreview(summary.userPreview);
-        marker.type = 'button';
-        marker.className = 'ai-session-conversation-marker';
-        marker.setAttribute('data-ai-session-conversation-marker', '');
-        marker.setAttribute('data-interaction-id', summary.id);
-        marker.setAttribute('data-response-state', summary.responseState);
-        marker.setAttribute('role', 'option');
-        var stroke = document.createElement('span');
-        var previewNode = document.createElement('span');
-        stroke.className = 'ai-session-conversation-marker-stroke';
-        stroke.setAttribute('aria-hidden', 'true');
-        previewNode.className = 'ai-session-conversation-marker-preview';
-        previewNode.textContent = preview;
-        marker.appendChild(stroke);
-        marker.appendChild(previewNode);
-        var label = getConversationTimestampLabel(summary.timestamp)
-            + ' — ' + preview;
-        marker.title = label;
-        marker.setAttribute('aria-label', label);
-        if (index === outline.interactions.length - 1) {
-            marker.setAttribute('data-latest', '');
-        }
-        if (summary.responseState === 'inProgress') {
-            marker.setAttribute('data-current', '');
-        }
-        marker.addEventListener('click', event => {
-            event.preventDefault();
-            event.stopPropagation();
-            focusConversationMarker(markers, index);
-            activateConversationMarker(marker, state, window.vscode);
-        });
-        marker.addEventListener('keydown', event => {
-            handleConversationMarkerKeydown(event, marker, state);
-        });
-        rail.appendChild(marker);
-    });
-
-    var markers = Array.from(rail.querySelectorAll(
-        '[data-ai-session-conversation-marker]'
-    ));
-    var selectedIndex = markers.findIndex(marker =>
-        marker.getAttribute('data-interaction-id') === selectedInteractionId
-    );
-    if (selectedIndex >= 0) {
-        setConversationMarkerSelection(markers, selectedIndex, false);
-    } else if (retainedFocusedInteractionId) {
-        markers.forEach(marker => {
-            marker.tabIndex = -1;
-            marker.setAttribute('aria-selected', 'false');
-        });
-    } else {
-        setConversationMarkerSelection(markers, markers.length - 1, false);
-    }
-
-    var hasOmittedInteractions = outline.partial
-        || outline.totalInteractions > outline.interactions.length;
-    count.textContent = hasOmittedInteractions
-        && outline.totalInteractions > ACTIVE_AI_SESSION_CONVERSATION_MAX_INTERACTIONS
-        ? '2,000+'
-        : String(outline.totalInteractions);
-    if (!markers.length) {
-        rail.hidden = true;
-        setActiveAiSessionConversationStatus(
-            row,
-            'empty',
-            'No user inputs yet'
-        );
-        state.hasRenderedOutline = true;
-        syncActiveAiSessionConversationListHeight(row);
-        return true;
-    }
-
-    rail.hidden = false;
-    if (hasOmittedInteractions) {
-        setActiveAiSessionConversationStatus(
-            row,
-            'partial',
-            'Older inputs omitted'
-        );
-    } else {
-        status.hidden = true;
-        status.removeAttribute('data-state');
-    }
-    syncActiveAiSessionConversationListHeight(row);
-
-    var retainedFocusedMarker = retainedFocusedInteractionId
-        ? markers.find(marker =>
-            marker.getAttribute('data-interaction-id')
-                === retainedFocusedInteractionId
-        )
-        : null;
-    var railAnchor = restoreState?.railAnchor || currentRailAnchor;
-    restoreConversationRailState(rail, railAnchor);
-    if (retainedFocusedMarker) {
-        focusConversationMarker(
-            markers,
-            markers.indexOf(retainedFocusedMarker)
-        );
-        restoreConversationRailState(rail, railAnchor);
-    }
-    state.hasRenderedOutline = true;
-    return true;
-}
-
-function renderActiveAiSessionConversationError(
-    row,
-    state,
-    error,
-    retainedRetryDeadline
-) {
-    clearActiveAiSessionConversationRetryTimer();
-    var panel = row && row.querySelector('[data-ai-session-conversation-panel]');
-    var rail = panel && panel.querySelector('[data-ai-session-conversation-rail]');
-    var count = panel && panel.querySelector('[data-ai-session-conversation-count]');
-    if (!panel || !rail || !count) return false;
-    clearConversationElement(rail);
-    rail.hidden = true;
-    count.textContent = '0';
-    state.sourceRevision = null;
-    state.interactionIds = new Set();
-    state.hasRenderedOutline = false;
-    state.restoreState = null;
-
-    var stateName = error.code === 'staleRevision' ? 'stale' : 'unavailable';
-    var text = error.code === 'staleRevision'
-        ? 'Conversation history changed'
-        : error.code === 'tooLarge'
-            ? 'Conversation history is too large'
-            : error.code === 'timeout'
-                ? 'Conversation history timed out'
-                : 'Conversation history unavailable';
-    var hint = '';
-    if (error.reason === 'reconnectingCodex') {
-        stateName = 'reconnecting';
-        text = 'Reconnecting to Codex…';
-    } else if (error.reason === 'codexRetryExhausted') {
-        text = 'Codex conversation history unavailable';
-    } else if (error.reason === 'updateCodex') {
-        text = 'Update Codex to view conversation history';
-    } else if (error.reason === 'unsupportedCodexProtocol') {
-        text = 'Installed Codex protocol is not supported';
-        hint = 'Compare your installed Codex and Agent Pivot versions';
-    }
-    setActiveAiSessionConversationStatus(
-        row,
-        stateName,
-        text,
-        hint,
-        error.retryAfterMs,
-        retainedRetryDeadline
-    );
-    state.validatedRender = {
-        kind: 'error',
-        error: error,
-        retryDeadline: activeAiSessionConversationRetryDeadline,
-    };
-    syncActiveAiSessionConversationListHeight(row);
-    return true;
-}
-
-function applyAiSessionConversationOutlineResult(message) {
-    var state = activeAiSessionConversationSubscription;
-    var hasPayload = isPlainConversationObject(message)
-        && Object.prototype.hasOwnProperty.call(message, 'payload');
-    var hasError = isPlainConversationObject(message)
-        && Object.prototype.hasOwnProperty.call(message, 'error');
-    var messageKeys = [
-        'type', 'version', 'requestId', 'subscriptionGeneration',
-        'projectId', 'provider', 'sessionId',
-    ];
-    if (!state
-        || !hasExactConversationKeys(
-            message,
-            messageKeys.concat(hasPayload ? ['payload'] : ['error']),
-            []
-        )
-        || hasPayload === hasError
-        || message.type !== 'ai-session-conversation-outline-result'
-        || message.version !== 1
-        || !Number.isSafeInteger(message.requestId)
-        || message.requestId < 1
-        || !Number.isSafeInteger(message.subscriptionGeneration)
-        || message.subscriptionGeneration < 0
-        || !isBoundedConversationIdentity(message.projectId)
-        || (message.provider !== 'codex'
-            && message.provider !== 'kimi'
-            && message.provider !== 'claude')
-        || !isBoundedConversationIdentity(message.sessionId)
-        || message.requestId !== state.outlineRequestId
-        || message.subscriptionGeneration !== state.subscriptionGeneration
-        || message.projectId !== state.projectId
-        || message.provider !== state.provider
-        || message.sessionId !== state.sessionId) {
-        return false;
-    }
-    var row = getCurrentActiveAiSessionConversationRow(state);
-    if (!row) return false;
-    if (hasPayload) {
-        if (!isValidConversationOutline(message.payload, state)) return false;
-        return renderActiveAiSessionConversationOutline(
-            row,
-            state,
-            message.payload
-        );
-    }
-    return isValidConversationError(message.error, state)
-        && renderActiveAiSessionConversationError(row, state, message.error);
-}
-
-function disconnectActiveAiSessionConversationResizeObserver() {
-    if (activeAiSessionConversationResizeObserver) {
-        activeAiSessionConversationResizeObserver.disconnect();
-        activeAiSessionConversationResizeObserver = null;
-    }
-    if (activeAiSessionConversationMutationObserver) {
-        activeAiSessionConversationMutationObserver.disconnect();
-        activeAiSessionConversationMutationObserver = null;
-    }
-}
-
-function getExpandedActiveAiSessionConversationRow() {
-    if (typeof document === 'undefined' || typeof document.querySelector !== 'function') {
-        return null;
-    }
-    return document.querySelector(
-        '.active-ai-session-row[data-conversation-expanded]'
-    );
-}
-
-function syncActiveAiSessionConversationListHeight(row) {
-    row = row || getExpandedActiveAiSessionConversationRow();
-    if (!row || !row.hasAttribute('data-conversation-expanded')) return false;
-    var panel = row.querySelector('[data-ai-session-conversation-panel]');
-    var conversationHeader = panel && panel.querySelector('header');
-    var list = row.closest('.codex-sessions-list');
-    if (!panel || !conversationHeader || !list) return false;
-
-    var collapsedRowHeight = Number(row.__stewardCollapsedConversationHeight);
-    if (!Number.isFinite(collapsedRowHeight) || collapsedRowHeight <= 0) {
-        collapsedRowHeight = Math.max(
-            0,
-            row.getBoundingClientRect().height - panel.getBoundingClientRect().height
-        );
-        row.__stewardCollapsedConversationHeight = collapsedRowHeight;
-    }
-    var collapsedListHeight = Number(list.__stewardCollapsedConversationHeight);
-    if (!Number.isFinite(collapsedListHeight) || collapsedListHeight <= 0) {
-        collapsedListHeight = list.getBoundingClientRect().height;
-        list.__stewardCollapsedConversationHeight = collapsedListHeight;
-    }
-
-    var panelStyle = typeof getComputedStyle === 'function'
-        ? getComputedStyle(panel)
-        : null;
-    var panelVerticalChrome = panelStyle
-        ? [
-            'marginTop', 'marginBottom',
-            'paddingTop', 'paddingBottom',
-            'borderTopWidth', 'borderBottomWidth',
-        ]
-            .reduce((total, property) =>
-                total + (parseFloat(panelStyle[property]) || 0), 0)
-        : 0;
-    var conversationHeaderHeight = conversationHeader.getBoundingClientRect().height;
-    var rail = panel.querySelector('[data-ai-session-conversation-rail]');
-    var autoScrollThreshold = getConversationAutoScrollThreshold(rail);
-    var railWasAtEnd = rail
-        && autoScrollThreshold !== null
-        && rail.scrollHeight - rail.clientHeight - rail.scrollTop
-            <= autoScrollThreshold;
-    row.style.removeProperty(
-        '--steward-ai-session-conversation-rail-height'
-    );
-    var visiblePanelContent = Array.from(panel.children).filter(child =>
-        child !== conversationHeader
-        && !child.hidden
-        && (typeof getComputedStyle !== 'function'
-            || getComputedStyle(child).display !== 'none')
-    );
-    var railScrollHeight = rail ? rail.scrollHeight || 0 : 0;
-    var railStyle = rail && typeof getComputedStyle === 'function'
-        ? getComputedStyle(rail)
-        : null;
-    var railMaxHeight = railStyle
-        ? parseFloat(railStyle.maxHeight)
-        : Number.NaN;
-    var naturalRailHeight = rail && visiblePanelContent.includes(rail)
-        ? Math.min(
-            railScrollHeight,
-            Number.isFinite(railMaxHeight) && railMaxHeight > 0
-                ? railMaxHeight
-                : railScrollHeight
-        )
-        : 0;
-    var naturalContentHeight = visiblePanelContent.reduce((total, child) =>
-        total + (child === rail
-            ? naturalRailHeight
-            : Math.max(
-                child.scrollHeight || 0,
-                child.getBoundingClientRect().height || 0
-            )), 0);
-    var naturalPanelHeight = conversationHeaderHeight
-        + naturalContentHeight
-        + panelVerticalChrome;
-    var naturalExpandedHeight = collapsedRowHeight + naturalPanelHeight;
-    var expansionDelta = Math.max(
-        0,
-        naturalExpandedHeight - collapsedRowHeight
-    );
-    var availableListHeight = Math.max(
-        collapsedRowHeight + conversationHeaderHeight + 72,
-        window.innerHeight - list.getBoundingClientRect().top - 8
-    );
-    var renderedListHeight = Math.min(
-        collapsedListHeight + expansionDelta,
-        availableListHeight
-    );
-    var railHeight = Math.max(
-        72,
-        renderedListHeight
-            - collapsedRowHeight
-            - conversationHeaderHeight
-            - panelVerticalChrome
-    );
-    list.style.setProperty(
-        '--steward-ai-session-expanded-extra-height',
-        Math.max(0, renderedListHeight - collapsedListHeight) + 'px'
-    );
-    if (naturalRailHeight > 0) {
-        row.style.setProperty(
-            '--steward-ai-session-conversation-rail-height',
-            Math.min(naturalRailHeight, railHeight) + 'px'
-        );
-        if (railWasAtEnd) {
-            rail.scrollTop = Math.max(0, rail.scrollHeight - rail.clientHeight);
-        }
-    }
-    return true;
-}
-
-function observeActiveAiSessionConversationSize(row) {
-    disconnectActiveAiSessionConversationResizeObserver();
-    var list = row && row.closest('.codex-sessions-list');
-    var panel = row && row.querySelector(
-        '[data-ai-session-conversation-panel]'
-    );
-    if (typeof ResizeObserver !== 'undefined' && row && list) {
-        activeAiSessionConversationResizeObserver = new ResizeObserver(() => {
-            syncActiveAiSessionConversationListHeight(row);
-        });
-        activeAiSessionConversationResizeObserver.observe(row);
-        activeAiSessionConversationResizeObserver.observe(list);
-    }
-    if (typeof MutationObserver !== 'undefined' && panel) {
-        activeAiSessionConversationMutationObserver = new MutationObserver(
-            () => syncActiveAiSessionConversationListHeight(row)
-        );
-        activeAiSessionConversationMutationObserver.observe(panel, {
-            attributes: true,
-            attributeFilter: ['hidden'],
-            childList: true,
-            subtree: true,
-        });
-    }
-    if (!activeAiSessionConversationResizeFallbackInstalled
-        && typeof window !== 'undefined'
-        && typeof window.addEventListener === 'function') {
-        window.addEventListener('resize', () => {
-            syncActiveAiSessionConversationListHeight();
-        });
-        activeAiSessionConversationResizeFallbackInstalled = true;
-    }
-}
-
-function applyActiveAiSessionConversationState(row, expanded, reveal) {
-    if (!row || typeof row.querySelector !== 'function') return false;
-    var panel = row.querySelector('[data-ai-session-conversation-panel]');
-    var header = row.querySelector('.ai-session-primary-action[aria-controls]');
-    var chevron = row.querySelector('.ai-session-conversation-chevron');
-    var list = row.closest('.codex-sessions-list');
-    if (!panel || !header || !chevron || !list) return false;
-
-    if (expanded) {
-        var otherRow = getExpandedActiveAiSessionConversationRow();
-        if (otherRow && otherRow !== row) {
-            applyActiveAiSessionConversationState(otherRow, false);
-        }
-        list.style.removeProperty('--steward-ai-session-expanded-extra-height');
-        list.__stewardCollapsedConversationHeight = list.getBoundingClientRect().height;
-        row.__stewardCollapsedConversationHeight = row.getBoundingClientRect().height;
-        row.setAttribute('data-conversation-expanded', '');
-        panel.hidden = false;
-        header.setAttribute('aria-expanded', 'true');
-        chevron.setAttribute('data-expanded', '');
-        expandedActiveAiSessionConversationKey = getActiveAiSessionConversationKey(
-            getActiveAiSessionConversationTarget(row)
-        );
-        syncActiveAiSessionConversationListHeight(row);
-        observeActiveAiSessionConversationSize(row);
-        if (expanded && reveal && typeof row.scrollIntoView === 'function') {
-            row.scrollIntoView({ block: 'nearest' });
-        }
-        return true;
-    }
-
-    row.removeAttribute('data-conversation-expanded');
-    panel.hidden = true;
-    header.setAttribute('aria-expanded', 'false');
-    chevron.removeAttribute('data-expanded');
-    row.style.removeProperty('--steward-ai-session-conversation-rail-height');
-    list.style.removeProperty('--steward-ai-session-expanded-extra-height');
-    list.__stewardCollapsedConversationHeight = null;
-    row.__stewardCollapsedConversationHeight = null;
-    if (expandedActiveAiSessionConversationKey === getActiveAiSessionConversationKey(
-        getActiveAiSessionConversationTarget(row)
-    )) {
-        expandedActiveAiSessionConversationKey = null;
-    }
-    prepareActiveAiSessionConversationLoading(row);
-    disconnectActiveAiSessionConversationResizeObserver();
-    return true;
-}
-
-function toggleActiveAiSessionConversation(row) {
-    var expanded = row && row.hasAttribute('data-conversation-expanded');
-    if (expanded) {
-        applyActiveAiSessionConversationState(row, false);
-        cancelActiveAiSessionConversation();
-        return false;
-    }
-
-    var previousRow = getExpandedActiveAiSessionConversationRow();
-    if (previousRow && previousRow !== row) {
-        var previousTarget = getActiveAiSessionConversationTarget(previousRow);
-        applyActiveAiSessionConversationState(previousRow, false);
-        cancelActiveAiSessionConversation(previousTarget);
-    }
-    if (!applyActiveAiSessionConversationState(row, true, true)) return false;
-    requestActiveAiSessionConversation(row);
-    return true;
-}
-
-function collapseActiveAiSessionConversation() {
-    var row = getExpandedActiveAiSessionConversationRow();
-    if (!row) return false;
-    var target = getActiveAiSessionConversationTarget(row);
-    applyActiveAiSessionConversationState(row, false);
-    cancelActiveAiSessionConversation(target);
-    return true;
-}
-
-function disposeActiveAiSessionConversation() {
-    clearActiveAiSessionConversationRetryTimer();
-    if (activeAiSessionConversationSubscription) {
-        activeAiSessionConversationSubscription.expanded = false;
-    }
-    activeAiSessionConversationSubscription = null;
-    expandedActiveAiSessionConversationKey = null;
-    disconnectActiveAiSessionConversationResizeObserver();
-}
-
-function captureExpandedConversationState(projectDiv) {
-    if (!projectDiv || typeof projectDiv.querySelector !== 'function') {
-        return null;
-    }
-    var row = projectDiv.querySelector(
-        '.active-ai-session-row[data-conversation-expanded]'
-    );
-    var rail = row?.querySelector('[data-ai-session-conversation-rail]');
-    var marker = rail?.querySelector('[data-ai-session-conversation-marker]:focus');
-    return row ? {
-        provider: row.getAttribute('data-session-provider') || '',
-        sessionId: row.getAttribute('data-session-id') || '',
-        expanded: true,
-        railAnchor: captureConversationRailState(rail),
-        focusedInteractionId: marker?.getAttribute('data-interaction-id') || '',
-    } : null;
-}
-
-function canRestoreExpandedConversation(projectDiv, state) {
-    if (!state?.expanded
-        || !projectDiv
-        || typeof projectDiv.querySelectorAll !== 'function') return null;
-    return Array.from(projectDiv.querySelectorAll(
-        '.active-ai-session-row[data-session-focused]'
-    )).find(row =>
-        row.getAttribute('data-session-provider') === state.provider
-        && row.getAttribute('data-session-id') === state.sessionId
-    ) || null;
-}
-
-function rebindActiveAiSessionConversation(row, capturedState) {
-    var target = getActiveAiSessionConversationTarget(row);
-    var subscription = activeAiSessionConversationSubscription;
-    if (!target || !subscription
-        || getActiveAiSessionConversationKey(target)
-            !== getActiveAiSessionConversationKey(subscription)) {
-        return false;
-    }
-    if (row.hasAttribute('data-conversation-expanded')
-        && !applyActiveAiSessionConversationState(row, false)) {
-        return false;
-    }
-    if (!applyActiveAiSessionConversationState(row, true, false)) return false;
-    subscription.expanded = true;
-    if (capturedState && (!subscription.restoreState
-        || subscription.validatedRender?.kind === 'outline')) {
-        subscription.restoreState = capturedState;
-    }
-    if (subscription.validatedRender?.kind === 'outline') {
-        renderActiveAiSessionConversationOutline(
-            row,
-            subscription,
-            subscription.validatedRender.outline
-        );
-    } else if (subscription.validatedRender?.kind === 'error') {
-        renderActiveAiSessionConversationError(
-            row,
-            subscription,
-            subscription.validatedRender.error,
-            subscription.validatedRender.retryDeadline
-        );
-    } else {
-        prepareActiveAiSessionConversationLoading(row);
-        syncActiveAiSessionConversationListHeight(row);
-    }
-    return true;
-}
-
-function captureCurrentWorkspaceConversationStates(root) {
-    var states = new Map();
-    if (!root || typeof root.querySelectorAll !== 'function') return states;
-    root.querySelectorAll(
-        '.workspace-card[data-current-workspace][data-id]'
-    ).forEach(projectDiv => {
-        var state = captureExpandedConversationState(projectDiv);
-        var projectId = projectDiv.getAttribute('data-id');
-        if (state && projectId) {
-            states.set(projectId, state);
-        }
-    });
-    if (states.size) {
-        disconnectActiveAiSessionConversationResizeObserver();
-    }
-    return states;
-}
-
-function restoreCurrentWorkspaceConversationStates(root, states) {
-    states = states || new Map();
-    if (!root || typeof root.querySelectorAll !== 'function') {
-        states.forEach((state, projectId) => {
-            cancelActiveAiSessionConversation({
-                projectId: projectId,
-                provider: state.provider,
-                sessionId: state.sessionId,
-            });
-        });
-        return;
-    }
-    root.querySelectorAll(
-        '.workspace-card[data-current-workspace][data-id]'
-    ).forEach(projectDiv => {
-        var projectId = projectDiv.getAttribute('data-id');
-        var state = states.get(projectId);
-        if (!state) return;
-        var row = canRestoreExpandedConversation(projectDiv, state);
-        if (!row || !rebindActiveAiSessionConversation(row, state)) {
-            cancelActiveAiSessionConversation({
-                projectId: projectId,
-                provider: state.provider,
-                sessionId: state.sessionId,
-            });
-        }
-        states.delete(projectId);
-    });
-    states.forEach((state, projectId) => {
-        cancelActiveAiSessionConversation({
-            projectId: projectId,
-            provider: state.provider,
-            sessionId: state.sessionId,
-        });
-    });
 }
 
 function getAdjacentAiSessionTab(tab, key) {
@@ -1470,12 +362,8 @@ function captureCurrentWorkspaceAiSessionStates(root) {
             states.set(projectId, {
                 view: captureAiSessionViewState(projectDiv),
                 providerMenu: captureAiSessionProviderMenuState(projectDiv),
-                conversation: captureExpandedConversationState(projectDiv),
             });
         });
-    if (Array.from(states.values()).some(state => state.conversation)) {
-        disconnectActiveAiSessionConversationResizeObserver();
-    }
     return states;
 }
 
@@ -1497,16 +385,6 @@ function restoreCurrentWorkspaceAiSessionViewStates(root, states, canRestoreProv
         });
 }
 
-function restoreCurrentWorkspaceAiSessionConversations(root, states) {
-    var conversationStates = new Map();
-    states.forEach((state, projectId) => {
-        if (state.conversation) {
-            conversationStates.set(projectId, state.conversation);
-        }
-    });
-    restoreCurrentWorkspaceConversationStates(root, conversationStates);
-}
-
 function restoreCurrentWorkspaceAiSessionAnchorsAndFocus(root, states) {
     if (!root || typeof root.querySelectorAll !== 'function') return;
     root.querySelectorAll('.workspace-card[data-current-workspace][data-id]')
@@ -1519,19 +397,6 @@ function restoreCurrentWorkspaceAiSessionAnchorsAndFocus(root, states) {
                 state.view.selectedTab,
                 { restoreFocus: false }
             );
-            var focusedConversationMarker = state.conversation
-                && state.conversation.focusedInteractionId
-                ? Array.from(projectDiv.querySelectorAll(
-                    '.active-ai-session-row[data-conversation-expanded] '
-                    + '[data-ai-session-conversation-marker]'
-                    + '[data-interaction-id]'
-                )).find(marker =>
-                    marker.getAttribute('data-interaction-id')
-                        === state.conversation.focusedInteractionId
-                    && document.activeElement === marker
-                )
-                : null;
-            if (focusedConversationMarker) return;
             restoreAiSessionViewFocus(projectDiv, state.view, selectedTab);
         });
 }
@@ -1594,7 +459,6 @@ function applyWorkspaceUpdate(message, options) {
             && typeof options.canRestoreAiSessionProviderMenu === 'function'
             && options.canRestoreAiSessionProviderMenu(projectId)
     );
-    restoreCurrentWorkspaceAiSessionConversations(replacement, aiSessionStates);
     restoreCurrentWorkspaceAiSessionAnchorsAndFocus(replacement, aiSessionStates);
     if (typeof window.__agentPivotSyncCollapseButton === 'function') {
         window.__agentPivotSyncCollapseButton();
@@ -1770,7 +634,6 @@ function applyOpenWorkspacesUpdate(message) {
             restoreAiSessionTabsFromState(document, window.vscode);
         }
         restoreCurrentWorkspaceAiSessionViewStates(wrapper, aiSessionStates);
-        restoreCurrentWorkspaceAiSessionConversations(wrapper, aiSessionStates);
         restoreCurrentWorkspaceAiSessionAnchorsAndFocus(wrapper, aiSessionStates);
         return false;
     }
@@ -1781,7 +644,6 @@ function applyOpenWorkspacesUpdate(message) {
         restoreAiSessionTabsFromState(document, window.vscode);
     }
     restoreCurrentWorkspaceAiSessionViewStates(wrapper, aiSessionStates);
-    restoreCurrentWorkspaceAiSessionConversations(wrapper, aiSessionStates);
     restoreCurrentWorkspaceAiSessionAnchorsAndFocus(wrapper, aiSessionStates);
     reconcilePendingOpenWorkspacePins(wrapper);
     var restoredPinButton = focusedPinButton
@@ -2109,7 +971,6 @@ function initProjects() {
     var pendingAiSessionProviderSelectionProjectId = null;
     var pendingAiSessionProviderSelectionRequestId = null;
     var pendingAiSessionProviderSelectionProviders = [];
-    window.addEventListener('pagehide', disposeActiveAiSessionConversation);
 
     function isDedicatedTodoTarget(target) {
         return Boolean(window.__agentPivotTodo
@@ -2492,13 +1353,6 @@ function initProjects() {
             && !activation.sessionRow.hasAttribute('data-session-pending')
             && activation.sessionRow.getAttribute('data-session-id')) {
             acknowledgeAiSessionRow(activation.sessionRow);
-        }
-        if (activation.toggleConversation) {
-            toggleActiveAiSessionConversation(activation.sessionRow);
-            return true;
-        }
-        if (activation.message?.type === 'focus-ai-session-terminal') {
-            collapseActiveAiSessionConversation();
         }
         if (activation.message) {
             window.vscode.postMessage(activation.message);
@@ -3624,11 +2478,6 @@ function initProjects() {
             focusAiSessionConversationOrigin(message);
             return;
         }
-        if (message
-            && message.type === 'ai-session-conversation-outline-result') {
-            applyAiSessionConversationOutlineResult(message);
-            return;
-        }
         if (message && message.type === 'todo-mutation-result') {
             applyTodoMutationResult(message, document);
             return;
@@ -4057,21 +2906,6 @@ function initProjects() {
             var tabProject = tab.closest('.project[data-id]');
             var tabProjectId = tabProject && tabProject.getAttribute('data-id');
             if (tabProjectId) onTriggerAiSessionAction(tab, tabProjectId);
-            return;
-        }
-
-        var expandedConversationRow = e.target && e.target.closest
-            ? e.target.closest(
-                '.active-ai-session-row[data-conversation-expanded]'
-            )
-            : null;
-        if (expandedConversationRow && e.key === 'Escape') {
-            e.preventDefault();
-            var conversationHeader = expandedConversationRow.querySelector(
-                '.ai-session-primary-action'
-            );
-            collapseActiveAiSessionConversation();
-            conversationHeader?.focus();
             return;
         }
