+++ tests/contract/aiSessions/conversationCoordinator.test.js
@@ -15,9 +15,6 @@ const { ConversationCoordinator } = require(
     '../../../out/aiSessions/conversation/coordinator'
 );
 const {
-    ConversationHostController,
-} = require('../../../out/aiSessions/conversation/conversationHostController');
-const {
     KimiConversationAdapter,
 } = require('../../../out/aiSessions/conversation/kimiAdapter');
 
@@ -76,32 +73,6 @@ function makePage(provider, sessionId, sourceRevision = 'native-a', overrides = 
         nextCursor: 'native-content-hash',
         isStart: false,
         isEnd: false,
-        ...overrides,
-    };
-}
-
-function makeOutlineRequest(overrides = {}) {
-    return {
-        type: 'request-ai-session-conversation-outline',
-        version: 1,
-        requestId: 1,
-        subscriptionGeneration: 1,
-        projectId: 'project-a',
-        provider: 'codex',
-        sessionId: 'session-a',
-        ...overrides,
-    };
-}
-
-function makeCancelRequest(overrides = {}) {
-    return {
-        type: 'cancel-ai-session-conversation',
-        version: 1,
-        requestId: 2,
-        subscriptionGeneration: 2,
-        projectId: 'project-a',
-        provider: 'codex',
-        sessionId: 'session-a',
         ...overrides,
     };
 }
@@ -236,77 +207,6 @@ function createCoordinatorHarness(adapterOverrides = {}) {
     return { adapters, calls, clock, coordinator, diagnostics };
 }
 
-function createControllerHarness(overrides = {}) {
-    const clock = createClock();
-    const calls = { codex: 0, kimi: 0, claude: 0 };
-    const outlineResults = [...(overrides.outlineResults || [])];
-    const adapters = {
-        codex: adapterReturning(calls, 'codex', {
-            readOutline: outlineResults.length
-                ? async sessionId => {
-                    const result = outlineResults.shift();
-                    return await result || makeOutline('codex', sessionId);
-                }
-                : overrides.readOutline,
-            readPage: overrides.readPage,
-        }),
-        kimi: adapterReturning(calls, 'kimi'),
-        claude: adapterReturning(calls, 'claude'),
-    };
-    const coordinator = new ConversationCoordinator({
-        adapters,
-        now: clock.now,
-        setTimeout: clock.setTimeout,
-        clearTimeout: clock.clearTimeout,
-        createCursorId: (() => {
-            let next = 0;
-            return () => `cursor-${++next}`;
-        })(),
-    });
-    const publications = [];
-    const opened = [];
-    let resolveCalls = 0;
-    const resolveTarget = overrides.resolveTarget || (
-        (projectId, provider, sessionId) => {
-            resolveCalls += 1;
-            return {
-                projectId,
-                provider,
-                sessionId,
-                focused: true,
-                executionState: 'running',
-                name: 'Conversation fixture',
-            };
-        }
-    );
-    const controller = new ConversationHostController({
-        coordinator,
-        resolveTarget,
-        publish: overrides.publish || (message => publications.push(message)),
-        openViewer: (target, authoritativeTarget) => {
-            opened.push([target, authoritativeTarget]);
-        },
-    });
-    return {
-        adapters,
-        calls,
-        clock,
-        controller,
-        coordinator,
-        get resolveCalls() {
-            return resolveCalls;
-        },
-        invalidate(provider = 'codex', sessionId = 'session-a') {
-            adapters[provider].invalidate(sessionId);
-        },
-        opened,
-        publications,
-        releaseSubscription(cancel = makeCancelRequest()) {
-            controller.cancel(cancel);
-        },
-    };
-}
-
 test('SESSION-CONVERSATION-COORDINATOR-001 isolates adapter failures and removes private details', async t => {
     const calls = { codex: 0, kimi: 0, claude: 0 };
     const coordinator = new ConversationCoordinator({
@@ -421,37 +321,6 @@ test('SESSION-CONVERSATION-COORDINATOR-003 projects stopped in-progress states w
     assert.equal(stopped.interactions.at(-1).responseState, 'interrupted');
 });
 
-test('SESSION-CONVERSATION-COORDINATOR-004 coalesces invalidations and applies the later debounce/rate-floor deadline', async t => {
-    const harness = createControllerHarness();
-    t.after(() => harness.controller.dispose());
-    await harness.controller.handleOutline(makeOutlineRequest());
-    harness.publications.length = 0;
-    const readsBeforeInvalidation = harness.calls.codex;
-
-    for (let index = 0; index < 10; index += 1) {
-        harness.invalidate();
-    }
-    harness.clock.advanceTo(249);
-    await settle();
-    assert.equal(harness.publications.length, 0);
-    harness.clock.advanceTo(250);
-    await settle();
-    assert.equal(harness.publications.length, 1);
-    assert.equal(harness.calls.codex, readsBeforeInvalidation + 1);
-
-    harness.clock.advanceTo(1050);
-    harness.invalidate();
-    harness.clock.advanceTo(1249);
-    await settle();
-    assert.equal(harness.publications.length, 1);
-    harness.clock.advanceTo(1299);
-    await settle();
-    assert.equal(harness.publications.length, 1);
-    harness.clock.advanceTo(1300);
-    await settle();
-    assert.equal(harness.publications.length, 2);
-});
-
 test('SESSION-CONVERSATION-COORDINATOR-005 releases only the exact opaque watch ownership', async () => {
     const { adapters, clock, coordinator } = createCoordinatorHarness();
     const calls = [];
@@ -480,578 +349,6 @@ test('SESSION-CONVERSATION-COORDINATOR-005 releases only the exact opaque watch 
     coordinator.dispose();
 });
 
-test('SESSION-CONVERSATION-HOST-001 rejects invalid request IDs before target resolution', async t => {
-    const harness = createControllerHarness();
-    t.after(() => harness.controller.dispose());
-
-    for (const requestId of [
-        0,
-        -1,
-        1.5,
-        Number.MAX_SAFE_INTEGER + 1,
-    ]) {
-        await harness.controller.handleOutline(makeOutlineRequest({ requestId }));
-    }
-    for (const subscriptionGeneration of [
-        -1,
-        1.5,
-        Number.MAX_SAFE_INTEGER + 1,
-    ]) {
-        await harness.controller.handleOutline(makeOutlineRequest({
-            subscriptionGeneration,
-        }));
-    }
-
-    assert.equal(harness.resolveCalls, 0);
-    assert.equal(harness.calls.codex, 0);
-    assert.deepEqual(harness.publications, []);
-});
-
-test('SESSION-CONVERSATION-HOST-002 fails unsupported identities closed without adapter access', async t => {
-    let resolveCalls = 0;
-    const harness = createControllerHarness({
-        resolveTarget: () => {
-            resolveCalls += 1;
-            return null;
-        },
-    });
-    t.after(() => harness.controller.dispose());
-
-    await harness.controller.handleOutline(makeOutlineRequest({
-        projectId: 'missing-project',
-    }));
-    await harness.controller.handleOutline(makeOutlineRequest({
-        sessionId: 'missing-session',
-        requestId: 2,
-    }));
-    await harness.controller.handleOutline(makeOutlineRequest({
-        provider: 'unsupported-provider',
-        requestId: 3,
-    }));
-
-    assert.equal(resolveCalls, 2);
-    assert.deepEqual(harness.calls, { codex: 0, kimi: 0, claude: 0 });
-    assert.deepEqual(
-        harness.publications.map(message => message.error?.code),
-        ['unavailable', 'unavailable']
-    );
-});
-
-test('SESSION-CONVERSATION-HOST-003 requires a target to remain focused through outline publication', async t => {
-    const pending = deferred();
-    let focused = true;
-    const harness = createControllerHarness({
-        outlineResults: [pending.promise],
-        resolveTarget: (projectId, provider, sessionId) => ({
-            projectId,
-            provider,
-            sessionId,
-            focused,
-            executionState: 'running',
-        }),
-    });
-    t.after(() => harness.controller.dispose());
-
-    const request = harness.controller.handleOutline(makeOutlineRequest());
-    focused = false;
-    pending.resolve(makeOutline('codex', 'session-a'));
-    await request;
-
-    assert.equal(harness.calls.codex, 1);
-    assert.deepEqual(harness.publications, []);
-});
-
-test('SESSION-CONVERSATION-HOST-004 suppresses request 9 after request 10 completes', async t => {
-    const first = deferred();
-    const second = deferred();
-    const published = [];
-    const harness = createControllerHarness({
-        outlineResults: [first.promise, second.promise],
-        publish: message => published.push(message),
-    });
-    t.after(() => harness.controller.dispose());
-
-    const request9 = harness.controller.handleOutline(makeOutlineRequest({
-        requestId: 9,
-        subscriptionGeneration: 3,
-    }));
-    const request10 = harness.controller.handleOutline(makeOutlineRequest({
-        requestId: 10,
-        subscriptionGeneration: 3,
-    }));
-    second.resolve(makeOutline('codex', 'session-a', 'native-b'));
-    await request10;
-    first.resolve(makeOutline('codex', 'session-a', 'native-a'));
-    await request9;
-
-    assert.deepEqual(published.map(message => message.requestId), [10]);
-});
-
-test('SESSION-CONVERSATION-HOST-005 suppresses generation 3 after generation 4 completes', async t => {
-    const first = deferred();
-    const second = deferred();
-    const published = [];
-    const harness = createControllerHarness({
-        outlineResults: [first.promise, second.promise],
-        publish: message => published.push(message),
-    });
-    t.after(() => harness.controller.dispose());
-
-    const generation3 = harness.controller.handleOutline(makeOutlineRequest({
-        requestId: 9,
-        subscriptionGeneration: 3,
-    }));
-    const generation4 = harness.controller.handleOutline(makeOutlineRequest({
-        requestId: 10,
-        subscriptionGeneration: 4,
-    }));
-    second.resolve(makeOutline('codex', 'session-a', 'native-b'));
-    await generation4;
-    first.resolve(makeOutline('codex', 'session-a', 'native-a'));
-    await generation3;
-
-    assert.deepEqual(published.map(message => [
-        message.requestId,
-        message.subscriptionGeneration,
-    ]), [[10, 4]]);
-});
-
-test('SESSION-CONVERSATION-HOST-006 cancels debounce and deferred publication on collapse', async t => {
-    const harness = createControllerHarness();
-    t.after(() => harness.controller.dispose());
-    await harness.controller.handleOutline(makeOutlineRequest({
-        subscriptionGeneration: 1,
-    }));
-    harness.publications.length = 0;
-    const readsBeforeRelease = harness.calls.codex;
-
-    harness.invalidate();
-    harness.releaseSubscription();
-    harness.releaseSubscription();
-    harness.clock.advanceBy(250);
-    await settle();
-
-    assert.equal(harness.calls.codex, readsBeforeRelease);
-    assert.equal(harness.publications.length, 0);
-});
-
-test('SESSION-CONVERSATION-HOST-007 ignores an old cancel for a newer exact subscription', async t => {
-    const harness = createControllerHarness();
-    t.after(() => harness.controller.dispose());
-    await harness.controller.handleOutline(makeOutlineRequest({
-        requestId: 10,
-        subscriptionGeneration: 4,
-    }));
-    harness.publications.length = 0;
-
-    harness.releaseSubscription(makeCancelRequest({
-        requestId: 9,
-        subscriptionGeneration: 3,
-    }));
-    harness.invalidate();
-    harness.clock.advanceBy(250);
-    await settle();
-
-    assert.equal(harness.publications.length, 1);
-});
-
-test('SESSION-CONVERSATION-HOST-008 cancel affects only the exact provider/session identity', async t => {
-    const harness = createControllerHarness();
-    t.after(() => harness.controller.dispose());
-    await harness.controller.handleOutline(makeOutlineRequest({
-        provider: 'codex',
-        sessionId: 'same-session',
-        requestId: 1,
-    }));
-    await harness.controller.handleOutline(makeOutlineRequest({
-        provider: 'kimi',
-        sessionId: 'same-session',
-        requestId: 2,
-    }));
-    harness.publications.length = 0;
-
-    harness.releaseSubscription(makeCancelRequest({
-        provider: 'codex',
-        sessionId: 'same-session',
-        requestId: 3,
-        subscriptionGeneration: 2,
-    }));
-    harness.invalidate('codex', 'same-session');
-    harness.invalidate('kimi', 'same-session');
-    harness.clock.advanceBy(250);
-    await settle();
-
-    assert.deepEqual(
-        harness.publications.map(message => message.provider),
-        ['kimi']
-    );
-});
-
-test('SESSION-CONVERSATION-HOST-009 hide and dispose release established watches and pending debounce timers', async () => {
-    for (const action of ['hide', 'dispose']) {
-        const harness = createControllerHarness();
-        await harness.controller.handleOutline(makeOutlineRequest());
-        harness.publications.length = 0;
-        const readsBeforeInvalidation = harness.calls.codex;
-        harness.invalidate();
-        assert.equal(harness.clock.pendingCount, 1);
-
-        if (action === 'hide') {
-            harness.controller.setVisible(false);
-        } else {
-            harness.controller.dispose();
-        }
-        assert.equal(harness.clock.pendingCount, 0);
-        assert.deepEqual(
-            harness.adapters.codex.disposedWatches,
-            ['session-a']
-        );
-        harness.clock.advanceBy(2_000);
-        await settle();
-
-        assert.equal(harness.calls.codex, readsBeforeInvalidation, action);
-        assert.deepEqual(harness.publications, [], action);
-        harness.controller.dispose();
-    }
-});
-
-test('SESSION-CONVERSATION-HOST-LIFECYCLE-001 reconcile republishes the exact running outline with stopped projection', async t => {
-    let executionState = 'running';
-    const harness = createControllerHarness({
-        resolveTarget: (projectId, provider, sessionId) => ({
-            projectId,
-            provider,
-            sessionId,
-            focused: true,
-            executionState,
-        }),
-    });
-    t.after(() => harness.controller.dispose());
-
-    await harness.controller.handleOutline(makeOutlineRequest());
-    assert.equal(
-        harness.publications.at(-1).payload.interactions.at(-1).responseState,
-        'inProgress'
-    );
-    const readsBeforeReconcile = harness.calls.codex;
-
-    executionState = 'stopped';
-    harness.controller.reconcile();
-    await settle();
-
-    assert.equal(harness.calls.codex, readsBeforeReconcile + 1);
-    assert.equal(harness.publications.length, 2);
-    assert.equal(
-        harness.publications.at(-1).payload.interactions.at(-1).responseState,
-        'interrupted'
-    );
-    assert.equal(
-        harness.publications.at(-1).subscriptionGeneration,
-        makeOutlineRequest().subscriptionGeneration
-    );
-});
-
-test('SESSION-CONVERSATION-HOST-LIFECYCLE-002 a later reconcile suppresses an older lifecycle read', async t => {
-    const stoppedRead = deferred();
-    const runningRead = deferred();
-    let executionState = 'running';
-    const harness = createControllerHarness({
-        outlineResults: [
-            makeOutline('codex', 'session-a'),
-            stoppedRead.promise,
-            runningRead.promise,
-        ],
-        resolveTarget: (projectId, provider, sessionId) => ({
-            projectId,
-            provider,
-            sessionId,
-            focused: true,
-            executionState,
-        }),
-    });
-    t.after(() => harness.controller.dispose());
-
-    await harness.controller.handleOutline(makeOutlineRequest());
-    harness.publications.length = 0;
-    executionState = 'stopped';
-    harness.controller.reconcile();
-    executionState = 'running';
-    harness.controller.reconcile();
-    runningRead.resolve(makeOutline('codex', 'session-a', 'native-running'));
-    await settle();
-    stoppedRead.resolve(makeOutline('codex', 'session-a', 'native-stopped'));
-    await settle();
-
-    assert.equal(harness.calls.codex, 3);
-    assert.equal(harness.publications.length, 1);
-    assert.equal(
-        harness.publications[0].payload.interactions.at(-1).responseState,
-        'inProgress'
-    );
-});
-
-test('SESSION-CONVERSATION-LOADING-001 coalesces repeated reconciliation behind the in-flight initial outline', async t => {
-    const pendingReads = [];
-    const harness = createControllerHarness({
-        readOutline: async (_sessionId) => {
-            const pending = deferred();
-            pendingReads.push(pending);
-            return pending.promise;
-        },
-    });
-    t.after(() => harness.controller.dispose());
-
-    const initialRequest = harness.controller.handleOutline(
-        makeOutlineRequest()
-    );
-    await settle();
-    assert.equal(pendingReads.length, 1);
-
-    for (let refresh = 0; refresh < 10; refresh += 1) {
-        harness.controller.reconcile();
-    }
-    await settle();
-
-    assert.equal(
-        pendingReads.length,
-        1,
-        'reconcile must not restart and starve the initial outline'
-    );
-    pendingReads[0].resolve(makeOutline(
-        'codex',
-        'session-a',
-        'native-initial'
-    ));
-    await initialRequest;
-    await settle();
-
-    assert.equal(harness.publications.length, 1);
-    assert.equal(
-        harness.publications[0].payload.sourceRevision,
-        'r1'
-    );
-    assert.equal(
-        pendingReads.length,
-        2,
-        'the reconciliations must coalesce into one follow-up read'
-    );
-
-    pendingReads[1].resolve(makeOutline(
-        'codex',
-        'session-a',
-        'native-follow-up'
-    ));
-    await settle();
-
-    assert.equal(harness.publications.length, 2);
-    assert.equal(
-        harness.publications[1].payload.sourceRevision,
-        'r2'
-    );
-});
-
-test('SESSION-CONVERSATION-HOST-LIFECYCLE-001 reconcile retries a missing state watch without duplicating recovery', async t => {
-    let watchAttempts = 0;
-    let activeInvalidation;
-    let outlineReads = 0;
-    let releaseCount = 0;
-    const publications = [];
-    const coordinator = {
-        setSessionStopped() {},
-        watch(_provider, _sessionId, onChange) {
-            watchAttempts += 1;
-            if (watchAttempts === 1) {
-                throw new Error('provider watch unavailable');
-            }
-            activeInvalidation = onChange;
-            return { id: 'recovered-watch' };
-        },
-        async readOutline(provider, sessionId) {
-            outlineReads += 1;
-            return makeOutline(provider, sessionId, `native-${outlineReads}`);
-        },
-        releaseSubscription(subscription) {
-            assert.equal(subscription.id, 'recovered-watch');
-            releaseCount += 1;
-            activeInvalidation = undefined;
-        },
-    };
-    const controller = new ConversationHostController({
-        coordinator,
-        resolveTarget: (projectId, provider, sessionId) => ({
-            projectId,
-            provider,
-            sessionId,
-            focused: true,
-            executionState: 'running',
-        }),
-        publish: message => publications.push(message),
-        openViewer: async () => undefined,
-    });
-    t.after(() => controller.dispose());
-
-    await controller.handleOutline(makeOutlineRequest());
-    assert.equal(watchAttempts, 1);
-    assert.equal(outlineReads, 1);
-    assert.equal(publications.length, 1);
-
-    controller.reconcile();
-    await settle();
-    assert.equal(watchAttempts, 2);
-    assert.equal(outlineReads, 2);
-    assert.equal(publications.length, 2);
-
-    controller.reconcile();
-    await settle();
-    assert.equal(watchAttempts, 2);
-    activeInvalidation();
-    await settle();
-    assert.equal(outlineReads, 4);
-    controller.dispose();
-    assert.equal(releaseCount, 1);
-});
-
-test('SESSION-CONVERSATION-HOST-010 opens only an authoritative known interaction', async t => {
-    const harness = createControllerHarness();
-    t.after(() => harness.controller.dispose());
-    await harness.controller.handleOutline(makeOutlineRequest());
-    const result = harness.publications[0];
-
-    await harness.controller.handleOpen({
-        type: 'open-ai-session-conversation',
-        version: 1,
-        requestId: 2,
-        subscriptionGeneration: 1,
-        projectId: 'project-a',
-        provider: 'codex',
-        sessionId: 'session-a',
-        interactionId: 'unknown',
-        expectedRevision: result.payload.sourceRevision,
-    });
-    await harness.controller.handleOpen({
-        type: 'open-ai-session-conversation',
-        version: 1,
-        requestId: 3,
-        subscriptionGeneration: 1,
-        projectId: 'project-a',
-        provider: 'codex',
-        sessionId: 'session-a',
-        interactionId: 'input-b',
-        expectedRevision: result.payload.sourceRevision,
-    });
-
-    assert.equal(harness.opened.length, 1);
-    assert.deepEqual(harness.opened[0][0], {
-        projectId: 'project-a',
-        provider: 'codex',
-        sessionId: 'session-a',
-        interactionId: 'input-b',
-        expectedRevision: 'r1',
-    });
-});
-
-test('SESSION-CONVERSATION-HOST-011 suppresses a stale adapter failure after focus changes', async t => {
-    const pending = deferred();
-    let focused = true;
-    const harness = createControllerHarness({
-        outlineResults: [pending.promise],
-        resolveTarget: (projectId, provider, sessionId) => ({
-            projectId,
-            provider,
-            sessionId,
-            focused,
-            executionState: 'running',
-        }),
-    });
-    t.after(() => harness.controller.dispose());
-
-    const request = harness.controller.handleOutline(makeOutlineRequest());
-    focused = false;
-    pending.reject(new Error('private failure after focus moved'));
-    await request;
-
-    assert.deepEqual(harness.publications, []);
-});
-
-test('SESSION-CONVERSATION-HOST-012 a newer cancel suppresses the exact deferred old publication', async t => {
-    const pending = deferred();
-    const harness = createControllerHarness({
-        outlineResults: [pending.promise],
-    });
-    t.after(() => harness.controller.dispose());
-
-    const request = harness.controller.handleOutline(makeOutlineRequest({
-        requestId: 8,
-        subscriptionGeneration: 3,
-    }));
-    harness.releaseSubscription(makeCancelRequest({
-        requestId: 9,
-        subscriptionGeneration: 4,
-    }));
-    pending.resolve(makeOutline('codex', 'session-a'));
-    await request;
-
-    assert.deepEqual(harness.publications, []);
-});
-
-test('SESSION-CONVERSATION-HOST-013 malformed cancel envelopes cannot release a current subscription', async t => {
-    const harness = createControllerHarness();
-    t.after(() => harness.controller.dispose());
-    await harness.controller.handleOutline(makeOutlineRequest({
-        requestId: 8,
-        subscriptionGeneration: 3,
-    }));
-    harness.publications.length = 0;
-
-    for (const requestId of [0, -1, 1.5, Number.MAX_SAFE_INTEGER + 1]) {
-        harness.releaseSubscription(makeCancelRequest({
-            requestId,
-            subscriptionGeneration: 4,
-        }));
-    }
-    harness.invalidate();
-    harness.clock.advanceBy(250);
-    await settle();
-
-    assert.equal(harness.publications.length, 1);
-});
-
-test('SESSION-CONVERSATION-COORDINATOR-001 a superseded late outline cannot invalidate its published revision', async t => {
-    const first = deferred();
-    const second = deferred();
-    const harness = createControllerHarness({
-        outlineResults: [first.promise, second.promise],
-        readPage: async request => makePage(
-            'codex',
-            request.sessionId,
-            'native-b'
-        ),
-    });
-    t.after(() => harness.controller.dispose());
-
-    const request9 = harness.controller.handleOutline(makeOutlineRequest({
-        requestId: 9,
-        subscriptionGeneration: 3,
-    }));
-    const request10 = harness.controller.handleOutline(makeOutlineRequest({
-        requestId: 10,
-        subscriptionGeneration: 4,
-    }));
-    second.resolve(makeOutline('codex', 'session-a', 'native-b'));
-    await request10;
-    const publishedRevision = harness.publications[0].payload.sourceRevision;
-    first.resolve(makeOutline('codex', 'session-a', 'native-a'));
-    await request9;
-
-    const page = await harness.coordinator.readPage({
-        provider: 'codex',
-        sessionId: 'session-a',
-        anchorInteractionId: 'input-a',
-        direction: 'around',
-        expectedRevision: publishedRevision,
-    });
-    assert.equal(page.sourceRevision, publishedRevision);
-});
 
 test('SESSION-CONVERSATION-COORDINATOR-001 accepts bounded long-prompt previews from every production adapter', async t => {
     const providerRoot = await fs.promises.mkdtemp(
@@ -1167,41 +464,6 @@ test('SESSION-CONVERSATION-COORDINATOR-001 accepts bounded long-prompt previews 
         ['claude', expectedPreview],
     ]);
 });
-
-test('SESSION-CONVERSATION-COORDINATOR-001 invalidation during initial publication triggers a second revision read', async t => {
-    const firstPublication = deferred();
-    const publications = [];
-    const harness = createControllerHarness({
-        outlineResults: [
-            makeOutline('codex', 'session-a', 'native-a'),
-            makeOutline('codex', 'session-a', 'native-b'),
-        ],
-        publish: message => {
-            publications.push(message);
-            return publications.length === 1
-                ? firstPublication.promise
-                : Promise.resolve();
-        },
-    });
-    t.after(() => harness.controller.dispose());
-
-    const request = harness.controller.handleOutline(makeOutlineRequest());
-    await settle();
-    assert.equal(publications.length, 1);
-    harness.invalidate();
-    harness.clock.advanceTo(250);
-    await settle();
-    firstPublication.resolve();
-    await request;
-    await settle();
-
-    assert.equal(harness.calls.codex, 2);
-    assert.deepEqual(
-        publications.map(message => message.payload.sourceRevision),
-        ['r1', 'r2']
-    );
-});
-
 test('SESSION-CONVERSATION-COORDINATOR-001 rate floor starts when a slow refresh publication completes', async t => {
     const firstPublication = deferred();
     const harness = createCoordinatorHarness();
